content service

+++ core/service/impl/content_service.go
@@ -34,23 +34,20 @@ func NewContentService(rep content.IContentRepo, q *query.ContentQuery) *content
 	}
 }
 
-
-
 // 获取页面
 func (c *contentService) GetPage(_ context.Context, id *proto.IdOrName) (*proto.SPage, error) {
 	ic := c._contentRepo.GetContent(0)
 	var ia content.IPage
-	if id.Id >0{
+	if id.Id > 0 {
 		ia = ic.GetPage(int32(id.Id))
-	}else{
+	} else {
 		ia = ic.GetPageByStringIndent(id.Name)
 	}
 	if ia != nil {
-		return c.parsePageDto(ia.GetValue()),nil
+		return c.parsePageDto(ia.GetValue()), nil
 	}
-	return nil,nil
+	return nil, nil
 }
-
 
 // 保存页面
 func (c *contentService) SavePage(_ context.Context, v *proto.SPage) (*proto.Result, error) {
@@ -65,19 +62,17 @@ func (c *contentService) SavePage(_ context.Context, v *proto.SPage) (*proto.Res
 	}
 	err = ip.SetValue(iv)
 	if err == nil {
-		_,err = ip.Save()
+		_, err = ip.Save()
 	}
-	return c.error(err),nil
+	return c.error(err), nil
 }
-
 
 // 删除页面
 func (c *contentService) DeletePage(_ context.Context, id *proto.Int64) (*proto.Result, error) {
 	ic := c._contentRepo.GetContent(0)
 	err := ic.DeletePage(int32(id.Value))
-	return c.error(err),nil
+	return c.error(err), nil
 }
-
 
 // 获取所有栏目
 func (c *contentService) GetArticleCategories(_ context.Context, empty *proto.Empty) (*proto.ArticleCategoriesResponse, error) {
@@ -88,25 +83,24 @@ func (c *contentService) GetArticleCategories(_ context.Context, empty *proto.Em
 		arr[i] = c.parseArticleCategoryDto(val)
 	}
 	return &proto.ArticleCategoriesResponse{
-		Value:arr,
-	},nil
+		Value: arr,
+	}, nil
 }
 
 // 获取栏目
 func (c *contentService) GetArticleCategory(_ context.Context, name *proto.IdOrName) (*proto.SArticleCategory, error) {
 	mgr := c._sysContent.ArticleManager()
 	var ic content.ICategory
-	if name.Id >0 {
+	if name.Id > 0 {
 		ic = mgr.GetCategory(int32(name.Id))
 	} else {
 		ic = mgr.GetCategoryByAlias(name.Name)
 	}
 	if c != nil {
-		return c.parseArticleCategoryDto(ic.GetValue()),nil
+		return c.parseArticleCategoryDto(ic.GetValue()), nil
 	}
-	return nil,nil
+	return nil, nil
 }
-
 
 // 保存文章栏目
 func (c *contentService) SaveArticleCategory(_ context.Context, r *proto.SArticleCategory) (*proto.Result, error) {
@@ -118,60 +112,55 @@ func (c *contentService) SaveArticleCategory(_ context.Context, r *proto.SArticl
 	}
 	err := ic.SetValue(v)
 	if err == nil {
-		_,err = ic.Save()
+		_, err = ic.Save()
 	}
-	return c.error(err),nil
+	return c.error(err), nil
 }
-
 
 // 删除文章分类
 func (c *contentService) DeleteArticleCategory(_ context.Context, id *proto.Int64) (*proto.Result, error) {
 	err := c._sysContent.ArticleManager().DelCategory(int32(id.Value))
-	return c.error(err),nil
+	return c.error(err), nil
 }
-
 
 // 获取文章
 func (c *contentService) GetArticle(_ context.Context, id *proto.IdOrName) (*proto.SArticle, error) {
-	m :=  c._sysContent.ArticleManager()
+	m := c._sysContent.ArticleManager()
 	var ia content.IArticle
-	if id.Id > 0{
+	if id.Id > 0 {
 		ia = m.GetArticle(int32(id.Id))
-	}else{
+	} else {
 
 	}
 	if ia != nil {
 		v := ia.GetValue()
-		return c.parseArticleDto(&v),nil
+		return c.parseArticleDto(&v), nil
 	}
-	return nil,nil
+	return nil, nil
 }
 
 // 删除文章
 func (c *contentService) DeleteArticle(_ context.Context, id *proto.Int64) (*proto.Result, error) {
 	err := c._sysContent.ArticleManager().DeleteArticle(int32(id.Value))
-	return c.error(err),nil
+	return c.error(err), nil
 }
-
 
 // 保存文章
 func (c *contentService) SaveArticle(_ context.Context, r *proto.SArticle) (*proto.Result, error) {
 	m := c._sysContent.ArticleManager()
 	v := c.parseArticle(r)
 	var ia content.IArticle
-	if r.Id > 0{
-		ia = m.GetArticle(r.Id)
-	}else{
+	if r.Id > 0 {
+		ia = m.GetArticle(int32(r.Id))
+	} else {
 		ia = m.CreateArticle(v)
 	}
 	err := ia.SetValue(v)
 	if err == nil {
-		_,err = ia.Save()
+		_, err = ia.Save()
 	}
-	return c.error(err),nil
+	return c.error(err), nil
 }
-
-
 
 func (c *contentService) QueryPagingArticles(_ context.Context, r *proto.PagingArticleRequest) (*proto.ArticleListResponse, error) {
 	var total = 0
@@ -187,16 +176,16 @@ func (c *contentService) QueryPagingArticles(_ context.Context, r *proto.PagingA
 	return &proto.ArticleListResponse{
 		Total: int64(total),
 		Data:  arr,
-	},nil
+	}, nil
 }
 
 func (c *contentService) QueryTopArticles(_ context.Context, cat *proto.IdOrName) (*proto.ArticleListResponse, error) {
 	var arr = make([]*proto.SArticle, 0)
 	var ic content.ICategory
 	var m = c._sysContent.ArticleManager()
-	if cat.Id > 0{
+	if cat.Id > 0 {
 		ic = m.GetCategory(int32(cat.Id))
-	}else {
+	} else {
 		ic = m.GetCategoryByAlias(cat.Name)
 	}
 	if ic != nil && ic.GetDomainId() > 0 {
@@ -208,32 +197,27 @@ func (c *contentService) QueryTopArticles(_ context.Context, cat *proto.IdOrName
 	return &proto.ArticleListResponse{
 		Total: 0,
 		Data:  arr,
-	},nil
+	}, nil
 }
-
-
-
-
-
 
 func (c *contentService) parsePageDto(src *content.Page) *proto.SPage {
 	return &proto.SPage{
-		Id:                   int64(src.Id),
-		UserId:               src.UserId,
-		Title:                src.Title,
-		StrIndent:            src.StrIndent,
-		PermFlag:             int32(src.PermFlag),
-		AccessKey:            src.AccessKey,
-		KeyWord:              src.KeyWord,
-		Description:          src.Description,
-		CssPath:              src.CssPath,
-		Body:                 src.Body,
-		UpdateTime:           src.UpdateTime,
-		Enabled:              src.Enabled == 1,
+		Id:          int64(src.Id),
+		UserId:      src.UserId,
+		Title:       src.Title,
+		StrIndent:   src.StrIndent,
+		PermFlag:    int32(src.PermFlag),
+		AccessKey:   src.AccessKey,
+		KeyWord:     src.KeyWord,
+		Description: src.Description,
+		CssPath:     src.CssPath,
+		Body:        src.Body,
+		UpdateTime:  src.UpdateTime,
+		Enabled:     src.Enabled == 1,
 	}
 }
 
-func (c *contentService) parsePage(v *proto.SPage) *content.Page{
+func (c *contentService) parsePage(v *proto.SPage) *content.Page {
 	return &content.Page{
 		Id:          int32(v.Id),
 		UserId:      v.UserId,
@@ -246,26 +230,26 @@ func (c *contentService) parsePage(v *proto.SPage) *content.Page{
 		CssPath:     v.CssPath,
 		Body:        v.Body,
 		UpdateTime:  v.UpdateTime,
-		Enabled:     types.IntCond(v.Enabled,1,0),
+		Enabled:     types.IntCond(v.Enabled, 1, 0),
 	}
 }
 
 func (c *contentService) parseArticleCategoryDto(v content.ArticleCategory) *proto.SArticleCategory {
 	return &proto.SArticleCategory{
-		Id:                   int64(v.ID),
-		ParentId:             int64(v.ParentId)	,
-		PermFlag:             int32(v.PermFlag),
-		Name:                 v.Name,
-		Alias:                v.Alias,
-		SortNum:              int32(v.SortNum),
-		Location:             v.Location,
-		Title:                v.Title,
-		Keywords:             v.Keywords,
-		Description:          v.Description,
+		Id:          int64(v.ID),
+		ParentId:    int64(v.ParentId),
+		PermFlag:    int32(v.PermFlag),
+		Name:        v.Name,
+		Alias:       v.Alias,
+		SortNum:     int32(v.SortNum),
+		Location:    v.Location,
+		Title:       v.Title,
+		Keywords:    v.Keywords,
+		Description: v.Description,
 	}
 }
 
-func (c *contentService) parseArticleCategory(r *proto.SArticleCategory) *content.ArticleCategory{
+func (c *contentService) parseArticleCategory(r *proto.SArticleCategory) *content.ArticleCategory {
 	return &content.ArticleCategory{
 		ID:          int32(r.Id),
 		ParentId:    int32(r.ParentId),
@@ -282,12 +266,12 @@ func (c *contentService) parseArticleCategory(r *proto.SArticleCategory) *conten
 
 func (c *contentService) parseArticleDto(v *content.Article) *proto.SArticle {
 	return &proto.SArticle{
-		Id:          v.ID,
-		CategoryId:       v.CatId,
+		Id:          int64(v.ID),
+		CategoryId:  int64(v.CatId),
 		Title:       v.Title,
 		SmallTitle:  v.SmallTitle,
 		Thumbnail:   v.Thumbnail,
-		PublisherId: int32(v.PublisherId),
+		PublisherId: int64(v.PublisherId),
 		Location:    v.Location,
 		Priority:    int32(v.Priority),
 		AccessKey:   v.AccessKey,
@@ -295,15 +279,15 @@ func (c *contentService) parseArticleDto(v *content.Article) *proto.SArticle {
 		Tags:        v.Tags,
 		ViewCount:   int32(v.ViewCount),
 		SortNum:     int32(v.SortNum),
-		CreateTime:  int32(v.CreateTime),
-		UpdateTime:  int32(v.UpdateTime),
+		CreateTime:  v.CreateTime,
+		UpdateTime:  v.UpdateTime,
 	}
 }
 
 func (c *contentService) parseArticle(v *proto.SArticle) *content.Article {
 	return &content.Article{
-		ID:          v.Id,
-		CatId:       v.CategoryId,
+		ID:          int32(v.Id),
+		CatId:       int32(v.CategoryId),
 		Title:       v.Title,
 		SmallTitle:  v.SmallTitle,
 		Thumbnail:   v.Thumbnail,
@@ -313,11 +297,9 @@ func (c *contentService) parseArticle(v *proto.SArticle) *content.Article {
 		PublisherId: int(v.PublisherId),
 		Content:     v.Content,
 		Tags:        v.Tags,
-		ViewCount:  int( v.ViewCount),
-		SortNum:    int( v.SortNum),
-		CreateTime: int64( v.CreateTime),
-		UpdateTime: int64( v.UpdateTime),
+		ViewCount:   int(v.ViewCount),
+		SortNum:     int(v.SortNum),
+		CreateTime:  v.CreateTime,
+		UpdateTime:  v.UpdateTime,
 	}
 }
-
-
